requirements.txt
Django>=4.2,<5.0
python-slugify>=8.0.0
